@@ -1,16 +1,27 @@
 {
-  "name": "rylac-ai-chat",
+  "name": "rylac",
   "version": "1.0.0",
-  "description": "RyLac AI Chat Application",
+  "description": "Instagram Downloader for photos, videos, and stories",
   "main": "server.js",
+  "type": "module",
   "scripts": {
     "start": "node server.js",
+    "dev": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "axios": "^1.6.2",
+    "cors": "^2.8.5",
+    "dotenv": "^16.3.1",
+    "multer": "^1.4.5-lts.1",
+    "file-type": "^18.0.0",
+    "form-data": "^4.0.0"
+  },
+  "devDependencies": {
+    "nodemon": "^3.0.2"
   },
   "engines": {
-    "node": ">=14.x"
+    "node": ">=16.x"
   }
 }
